Refactor validators.
