+++ lib/validators/index.js
@@ -1,0 +1,11 @@
+/*!
+ * Copyright (c) 2017-2019 Digital Bazaar, Inc. All rights reserved.
+ */
+'use strict';
+
+module.exports = {
+  createWebLedgerRecord: require('./createWebLedgerRecord'),
+  proofs: require('./proofs'),
+  updateWebLedgerRecord: require('./updateWebLedgerRecord'),
+  webLedgerConfiguration: require('./webLedgerConfiguration'),
+};
